Move code to file pkg
